requirements.txt
Django==3.0
django-tabbed-admin==1.0.4
Jinja2==3.0.3